tests/wpt/tests/css/css-flexbox/parsing/order-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing order with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-order">
<meta name="assert" content="order supports only the grammar '<integer>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("order", "auto");
test_invalid_value("order", "123.45");
test_invalid_value("order", "123 45");
</script>
</body>
</html>
